delineate between old deprecated fields and ones we intend to continue using.
